RABSW-781: Run ClientMount controller in kind
This commit runs the ClientMount controller in the DWS controller manager
when running in the kind environment. I added the OVERLAY variable to the
Makefile so kustomize can add environment specific patches. There are
currently patches for the kind, craystack, and dp0 environments. Each of them
set an environment variable in the controller manager's pod to specify the
environment.